+++ src/mapkey.cpp
@@ -133,11 +133,10 @@ std::string MapKey::argument_value(BPFtrace &bpftrace,
       std::vector<std::string> elems;
       for (auto &field : arg.GetStructFields())
       {
-        elems.push_back(
-            "." + field.first + "=" +
-            argument_value(bpftrace,
-                           field.second.type,
-                           (const uint8_t *)data + field.second.offset));
+        elems.push_back("." + field.name + "=" +
+                        argument_value(bpftrace,
+                                       field.type,
+                                       (const uint8_t *)data + field.offset));
       }
       return "{" + str_join(elems, ",") + "}";
     }